requirements.txt
torch
torchvision
pytorch-lightning
albumentations
opencv-python
scikit-image
Pillow